refactor(type_checker): extract methods/ with string receiver split
